clue_game_backend/clue_game_api/room/serializers.py
from rest_framework import serializers
from .models import Room


class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = "__all__"  # You can specify specific fields if needed
