clue_game_backend/requirements.txt
Django==3.2.*
dj-database-url==0.5.0
djangorestframework==3.13.*
psutil==5.6.*
psycopg2-binary==2.8.*
django-cors-headers==4.3.*